+++ CHANGELOG.md
@@ -9,9 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Added
 - Examples
 - Proper documentation.
+- `Drop` protocol and `drop` function.
+- `RefMut` type and `as_mut`.
 
 ## Changed
 - The `Some` type `__str__` now returns `Some(repr(value))` instead of just `repr(value)`.
+- `Ref` now supports `__hash__`.
 
 ## Removed
 - `Some.is_none_and` method since it made no sense.